refactor: update consumer imports to canonical module paths
Replace re-export-dependent import paths with canonical paths:

- homeboy::api → homeboy::server::api
- homeboy::auth → homeboy::server::auth
- homeboy::cli_tool → homeboy::engine::cli_tool
- crate::http → crate::server::http / super::http
- crate::hooks → crate::engine::hooks
- crate::permissions → crate::deploy::permissions / super::permissions
- homeboy::codebase_map → homeboy::code_audit::codebase_map
- homeboy::docs (inline refs) → homeboy::code_audit::docs

Public re-exports remain in core/mod.rs for external API stability.
Internal pub(crate) re-exports for http and permissions removed
since all internal consumers now use canonical paths.

776 tests pass, cargo fmt clean, zero new warnings.
