+++ src/core/extension/execution.rs
@@ -3,9 +3,9 @@ use crate::engine::command::CapturedOutput;
 use crate::engine::shell;
 use crate::engine::{template, validation};
 use crate::error::{Error, Result};
-use crate::http::ApiClient;
 use crate::local_files;
 use crate::project::{self, Project};
+use crate::server::http::ApiClient;
 use crate::server::{
     execute_local_command_in_dir, execute_local_command_interactive,
     execute_local_command_passthrough, CommandOutput,
